@@ -4,39 +4,31 @@
     "author": "NobuhitoArai <nobht.arai@gmail.com>",
     "private": true,
     "scripts": {
-        "start": "snowpack dev",
-        "build": "snowpack build",
-        "test": "web-test-runner \"src/**/*.test.ts\"",
-        "format": "prettier --write **/*.{js,ts,json,md,html,svelte}"
-    },
-    "prettier": {
-        "singleQuote": true,
-        "semi": false,
-        "tabWidth": 4,
-        "arrowParens": "avoid",
-        "svelteBracketNewLine": true,
-        "svelteStrictMode": true
-    },
-    "dependencies": {
-        "prettier": "^2.5.1",
-        "svelte": "^3.44.3",
-        "uikit": "^3.9.4"
+        "dev": "vite dev",
+        "build": "vite build",
+        "preview": "vite preview",
+        "check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
+        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
+        "lint": "prettier --plugin-search-dir . --check . && eslint .",
+        "format": "prettier --plugin-search-dir . --write ."
     },
     "devDependencies": {
-        "@snowpack/plugin-dotenv": "^2.2.0",
-        "@snowpack/plugin-optimize": "^0.2.13",
-        "@snowpack/plugin-sass": "^1.4.0",
-        "@snowpack/plugin-svelte": "^3.7.0",
-        "@snowpack/plugin-typescript": "^1.2.1",
-        "@snowpack/web-test-runner-plugin": "^0.2.2",
-        "@testing-library/svelte": "^3.0.3",
-        "@types/snowpack-env": "^2.3.4",
-        "@web/test-runner": "^0.13.22",
-        "chai": "^4.3.4",
-        "snowpack": "^3.8.8",
-        "svelte-preprocess": "^4.10.1",
-        "svelte-spa-router": "^3.2.0",
-        "typescript": "^4.5.4"
+        "@sveltejs/adapter-auto": "^2.0.0",
+        "@sveltejs/kit": "^1.20.4",
+        "eslint": "^8.28.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-svelte": "^2.30.0",
+        "prettier": "^2.8.0",
+        "prettier-plugin-svelte": "^2.10.1",
+        "svelte": "^4.0.5",
+        "svelte-check": "^3.4.3",
+        "typescript": "^5.0.0",
+        "uikit": "^3.16.22",
+        "vite": "^4.4.2"
     },
-    "packageManager": "yarn@3.1.1"
+    "packageManager": "yarn@3.1.1",
+    "type": "module",
+    "dependencies": {
+        "sass": "^1.63.6"
+    }
 }
