@@ -1,9 +1,16 @@
 <script lang="ts">
-
 </script>
+
+<footer class="footer">
+    <p class="footer__copyright">
+        <small>Copyright (c) NaoNao All Rights Reserved.</small>
+    </p>
+</footer>
 
 <style lang="scss">
     .footer {
+        position: sticky;
+        top: 100vh;
         padding: 1rem 0;
         background: #e2e2e2;
     }
@@ -15,9 +22,3 @@
         text-align: center;
     }
 </style>
-
-<footer class="footer">
-    <p class="footer__copyright">
-        <small>Copyright (c) NaoNao All Rights Reserved.</small>
-    </p>
-</footer>
