@@ -1,22 +1,5 @@
 <script lang="ts">
-
 </script>
-
-<style lang="scss">
-    .home {
-        width: 80%;
-        max-width: 1440px;
-        margin: auto;
-    }
-
-    .home__content {
-        display: block;
-
-        &:not(:first-child) {
-            margin-top: 3rem;
-        }
-    }
-</style>
 
 <section class="home">
     <h2>
@@ -43,10 +26,27 @@
         </p>
         <ul>
             <li><a href="https://svelte.dev/">Svelte</a></li>
-            <li><a href="https://www.snowpack.dev">Snowpack</a></li>
+            <li><a href="https://kit.svelte.jp/">Svelte Kit</a></li>
+            <li><a href="https://vitejs.dev">Vite</a></li>
             <li><a href="https://www.typescriptlang.org/">TypeScript</a></li>
             <li><a href="https://sass-lang.com/">Sass</a></li>
             <li><a href="https://getuikit.com">UIkit</a></li>
         </ul>
     </article>
 </section>
+
+<style lang="scss">
+    .home {
+        width: 80%;
+        max-width: 1440px;
+        margin: auto;
+    }
+
+    .home__content {
+        display: block;
+
+        &:not(:first-child) {
+            margin-top: 3rem;
+        }
+    }
+</style>
